Changed require statements to recursively find all rule .js files
