+++ lib/rules/ec2/aws-ec2-ami-includes-sql-server.spec.js
