+++ lib/rules/ec2/aws-ec2-has-key-pair.js
