+++ lib/rules/ec2/aws-ec2-key-pair-exists.js
