+++ lib/rules/ec2/aws-ec2-uses-key-pair.spec.js
