+++ lib/rules/ec2/aws-ec2-uses-security-group.spec.js
