+++ lib/rules/rds/aws-rds-encryption-key-exists.spec.js
