+++ lib/rules/security-group/aws-security-group-allows-inbound.js
@@ -2,8 +2,8 @@
 const co = require('co')
 const debug = require('debug')('tfrules/aws-security-group-allow-inbound');
 const _ = require('lodash');
-const cidrComparison = require('./modules/cidr-comparison');
-const portComparison = require('./modules/port-comparison');
+const cidrComparison = require('../modules/cidr-comparison');
+const portComparison = require('../modules/port-comparison');
 
 //------------------------------------------------------------------------------
 // Rule Definition
