+++ lib/rules/security-group/aws-security-group-allows-outbound.spec.js
