+++ lib/rules/security-group/aws-security-group-does-not-allow-outbound.js
@@ -1,22 +1,22 @@
 'use strict';
 const co = require('co')
-const debug = require('debug')('tfrules/aws-security-group-does-not-allow-inbound');
+const debug = require('debug')('tfrules/aws-security-group-does-not-allow-outbound');
 const _ = require('lodash');
-const cidrComparison = require('./modules/cidr-comparison');
-const portComparison = require('./modules/port-comparison');
+const cidrComparison = require('../modules/cidr-comparison');
+const portComparison = require('../modules/port-comparison');
 
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
 
-const AWSSecurityGroupDoesNotAllowInbound = {};
+const AWSSecurityGroupDoesNotAllowOutbound = {};
 
-AWSSecurityGroupDoesNotAllowInbound.docs = {
-  description: 'Security Group does not allow specified inbound traffic',
+AWSSecurityGroupDoesNotAllowOutbound.docs = {
+  description: 'Security Group does not allow specified outbound traffic',
   recommended: true
 };
 
-AWSSecurityGroupDoesNotAllowInbound.schema = {
+AWSSecurityGroupDoesNotAllowOutbound.schema = {
   type : 'object',
   properties : {
     cidr : {
@@ -26,7 +26,7 @@ AWSSecurityGroupDoesNotAllowInbound.schema = {
     port : {
       anyOf : [
         {
-          type: 'integer',
+          type : 'integer'
         },
         { 
           type : 'string',
@@ -41,11 +41,11 @@ AWSSecurityGroupDoesNotAllowInbound.schema = {
   }
 };
 
-AWSSecurityGroupDoesNotAllowInbound.paths = {
+AWSSecurityGroupDoesNotAllowOutbound.paths = {
   awsSecurityGroup : 'aws_security_group'
 };
 
-AWSSecurityGroupDoesNotAllowInbound.validate = function *( context ) {
+AWSSecurityGroupDoesNotAllowOutbound.validate = function *( context ) {
   // debug( '%O', context );
   let result = null;
   let message = [];
@@ -53,26 +53,24 @@ AWSSecurityGroupDoesNotAllowInbound.validate = function *( context ) {
   if( context.config ) {
     // debug('Config: %j', context.config)
     // debug('Instance: %j', context.instance)
-    for (let ingress of context.instance.ingress) {
-      // debug('Ingress: %j', ingress)
-      // debug('Send Groups: %j', groups)
+    for (let egress of context.instance.egress) {
+      // debug('Egress: %j', egress)
       if (context.config.cidr) {
-        groups[context.instance.name] = yield cidrComparison(context.config, ingress, false)
+        groups[context.instance.name] = yield cidrComparison(context.config, egress, false)
       } else if (context.config.port) {
-        groups[context.instance.name] = yield portComparison(context.config, ingress, false)
+        groups[context.instance.name] = yield portComparison(context.config, egress, false)
       }
     }
 
     debug('Insecure Groups: %j', groups)
     if ( ! _.isEmpty(groups[context.instance.name]) ){
-      debug('Insecure Groups: %j', groups)
       _.map(groups[context.instance.name], group => {
         if ( group.cidr && ! group.ports ) {
-          message.push('CIDR Block ' + group.cidr + ' in security group ' + context.instance.name + ' should not be open to inbound traffic from ips: ' + context.config.cidr)
+          message.push('CIDR Block ' + group.cidr + ' in security group ' + context.instance.name + ' should not be open to outbound traffic to ips: ' + context.config.cidr)
         } else if ( group.ports && ! group.cidr ) {
-          message.push('Security group ' + context.instance.name + ' should not be taking traffic on ports: ' + context.config.port)
+          message.push('Security group ' + context.instance.name + ' should not be sending traffic on ports: ' + context.config.port)
         } else if ( group.cidr && group.ports ) {
-          message.push('CIDR Block ' + group.cidr + ' in security group ' + context.instance.name + ' should not be taking inbound traffic on ports: ' + context.config.port + ' from ips: ' + context.config.cidr)
+          message.push('CIDR Block ' + group.cidr + ' in security group ' + context.instance.name + ' should not be sending outbound traffic on ports: ' + context.config.port + ' to ips: ' + context.config.cidr)
         }
       })
       result = {
@@ -89,5 +87,5 @@ AWSSecurityGroupDoesNotAllowInbound.validate = function *( context ) {
   return result;
 };
 
-module.exports = AWSSecurityGroupDoesNotAllowInbound;
+module.exports = AWSSecurityGroupDoesNotAllowOutbound;
