+++ lib/rules/security-group/aws-security-group-does-not-allow-outbound.spec.js
