+++ lib/rules/security-group/aws-security-group-exists.spec.js
